Module4 jpa

+++ src/main/java/com/belhard/bookstore/dao/BookDao.java
@@ -20,5 +20,5 @@ public interface BookDao {
 
     boolean deleteBook(Long id);
 
-    int countAllBooks();
+    Long countAllBooks();
 }
