+++ src/main/java/com/belhard/bookstore/dao/UserDao.java
@@ -20,5 +20,5 @@ public interface UserDao {
 
     boolean deleteUser(Long id);
 
-    int countAllUsers();
+    Long countAllUsers();
 }
